pyproject.toml
[project]
name = "baretrader"
version = "1.0.0"
description = "CLI-based automated trading system"
authors = [
    {name = "Joe Caruso"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "alpaca-py>=0.21.0",
    "rich>=13.0.0",
    "pytz>=2024.1",
    "pyarrow>=19.0.0",
    "bokeh>=3.3.0",
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
baretrader = "baretrader.cli.main:cli"

[tool.poetry]
packages = [{include = "trader"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
types-requests = "^2.31.0"
types-pyyaml = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # line length; many lines slightly over 120

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
